requirements-dev.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
alembic==1.6.2
amqp==2.6.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
arrow==0.17.0
attrs==21.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
authlib==0.15.3
b2sdk==1.7.0
babel==2.9.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
billiard==3.6.4.0
bleach==3.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
boto3==1.16.47
botocore==1.19.47
cachetools==4.2.2
celery==4.4.7
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
click==8.0.0; python_version >= '3.6'
clickclick==20.10.2
colorama==0.4.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
connexion==2.7.0
coverage[toml]==5.5
cryptography==3.3.2
decorator==5.0.7; python_version >= '3.5'
docutils==0.17.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dollar-ref==0.1.3
fakeredis==1.5.0
flake8==3.9.1
flask-babel==2.0.0
flask-cors==3.0.10
flask==1.1.2
funcsigs==1.0.2
future==0.18.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
geocoder==1.38.1
httpretty==1.1.1
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
importlib-metadata==4.0.1; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
iniconfig==1.1.1
isodate==0.6.0
itsdangerous==1.1.0
jeepney==0.6.0; sys_platform == 'linux'
jinja2==3.0.0; python_version >= '3.6'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
jsonschema==3.2.0
keyring==23.0.1; python_version >= '3.6'
kombu==4.6.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
ldap3==2.9
logfury==0.1.2
mako==1.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markupsafe==2.0.0; python_version >= '3.6'
mccabe==0.6.1
openapi-schema-validator==0.1.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
openapi-spec-validator==0.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
packaging==20.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
passlib==1.7.4
pkginfo==1.7.0
pluggy==0.13.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
py==1.10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyasn1==0.4.8
pycodestyle==2.7.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pycryptodomex==3.9.9
pyflakes==2.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pygments==2.9.0; python_version >= '3.5'
pyjwt==2.1.0
pymysql==1.0.2
pyotp==2.6.0
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyrsistent==0.17.3; python_version >= '3.5'
pytest-cov==2.12.0
pytest==6.2.4
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-editor==1.0.4
pytz==2021.1
pyyaml==5.4.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
ratelim==0.1.6
readme-renderer==29.0
redis==3.5.3
requests-toolbelt==0.9.1
requests==2.25.1
rfc3986==1.5.0
s3transfer==0.3.7
secretstorage==3.3.1; sys_platform == 'linux'
six==1.15.0
sortedcontainers==2.3.0
sqlalchemy-utils==0.37.0
sqlalchemy==1.3.24
swagger-ui-bundle==0.0.8
termcolor==1.1.0
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
tqdm==4.60.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
twine==3.4.1
urllib3==1.26.4
vine==1.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
webencodings==0.5.1
werkzeug==2.0.0; python_version >= '3.6'
wheel==0.36.2
zipp==3.4.1; python_version >= '3.6'
